python_legacy/requirements.txt
streamlit
Pillow
google-generativeai
python-dotenv
requests
PyPDF2
pydantic
# Office file support
python-docx
openpyxl
python-pptx
# Backend dependencies
fastapi
uvicorn
pubchempy
python-multipart
# Free OCR alternatives
pix2text